feat: Complete document service refactoring and logging migration
